finish navi bar
